Move sampler classes to dedicated package
