+++ src/main/java/com/example/model/llama/Llama.java
@@ -3,8 +3,8 @@ package com.example.model.llama;
 import com.example.auxiliary.Parallel;
 import com.example.auxiliary.format.LlamaChatFormat;
 import com.example.core.model.tensor.FloatTensor;
-import com.example.inference.Sampler;
 import com.example.model.Configuration;
+import com.example.inference.sampler.Sampler;
 import com.example.model.Model;
 import com.example.Options;
 import com.example.loader.weights.ModelLoader;
