+++ src/main/java/com/example/model/mistral/Mistral.java
@@ -3,8 +3,8 @@ package com.example.model.mistral;
 import com.example.auxiliary.Parallel;
 import com.example.auxiliary.format.MistralChatFormat;
 import com.example.core.model.tensor.FloatTensor;
-import com.example.inference.Sampler;
 import com.example.model.Configuration;
+import com.example.inference.sampler.Sampler;
 import com.example.model.Model;
 import com.example.Options;
 import com.example.loader.weights.ModelLoader;
